requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
mysql-connector-python==8.3.0 